Split scheduled endpoints into consent controllers
